+++ tests/test-methods.ts
@@ -7,13 +7,16 @@ import { Controller, Get, Post, Put, Del, AdvancedController } from '../dist/ind
 import { app, baseUrl } from './test-base';
 
 
-var localBaseUrl = baseUrl + 'methods/';
+var localBaseUrl = baseUrl + 'methods';
 
 
 @Controller('methods')
 class MethodTestController extends AdvancedController {
 	@Get()
 	get() { return 'get'; }
+
+	@Get('')
+	getEmpty() { return 'get-empty'; }
 
 	@Post()
 	post() { return 'post'; }
@@ -40,15 +43,23 @@ describe('MethodTestController', () => {
 	});
 
 	it('should return get', (done) => {
-		request.get(localBaseUrl + 'get', (err, res, body) => {
+		request.get(localBaseUrl + '/get', (err, res, body) => {
 			let data = assertAndParse(err, res, body);
 			assert.equal(data, 'get');
 			done();
 		});
 	});
 
+	it('should handle empty action name', (done) => {
+		request.get(localBaseUrl + '', (err, res, body) => {
+			let data = assertAndParse(err, res, body);
+			assert.equal(data, 'get-empty');
+			done();
+		});
+	});
+
 	it('should return post', (done) => {
-		request.post(localBaseUrl + 'post', (err, res, body) => {
+		request.post(localBaseUrl + '/post', (err, res, body) => {
 			let data = assertAndParse(err, res, body);
 			assert.equal(data, 'post');
 			done();
@@ -56,7 +67,7 @@ describe('MethodTestController', () => {
 	});
 
 	it('should return put', (done) => {
-		request.put(localBaseUrl + 'put', (err, res, body) => {
+		request.put(localBaseUrl + '/put', (err, res, body) => {
 			let data = assertAndParse(err, res, body);
 			assert.equal(data, 'put');
 			done();
@@ -64,7 +75,7 @@ describe('MethodTestController', () => {
 	});
 
 	it('should return delete (use @del, not @delete)', (done) => {
-		request.del(localBaseUrl + 'delete', (err, res, body) => {
+		request.del(localBaseUrl + '/delete', (err, res, body) => {
 			let data = assertAndParse(err, res, body);
 			assert.equal(data, 'delete');
 			done();